refactor(auto-topup): move EIP-712 verifier into lit-billing-core
Glitch's PR #448 review flagged that lit-payments hits lit-api-server
over HTTP just to verify a wallet signature when the verifier is pure
Rust with no on-chain dependency. Folding it into the shared crate lets
both services call it in-process.

What moved:
- `lit-api-server/src/core/eip712.rs` (full implementation, ~25 tests)
  → `lit-billing-core/src/eip712.rs`
- `verify_eip712_signature` now takes `chain_id: u64` as a parameter
  instead of reading lit-api-server's `GLOBAL_NODE_CONFIG`; each service
  supplies its own chain_id.
- Replaces lit-api-server-specific `ApiStatus` returns with a new
  `Eip712Error` enum (BadRequest / Internal) so the shared crate has no
  knowledge of Rocket response types.

Adapter on lit-api-server side: `core::eip712` is now a thin wrapper
that reads chain_id from GLOBAL_NODE_CONFIG and translates Eip712Error
back to ApiStatus, so existing callers in `account_management.rs` keep
their signatures unchanged.

Cites glitch's comment on `lit-api-server/src/internal/routes.rs:90`
(/internal/verify_wallet_auth should not exist).
