+++ lit-billing-core/src/eip712.rs
@@ -1,0 +1,1037 @@
+//! EIP-712 typed-data verification for ChainSecured signatures (CPL-286).
+//!
+//! Wallet UIs render the signed payload as a typed struct with `primaryType`
+//! clearly labeled. Cross-flow replay is rejected at the type-hash level:
+//! a signature minted against `CreateWallet` cannot recover when the server
+//! hashes the same bytes against, say, `BillingAuth` or `AddUsageApiKey` —
+//! the type hash commits to the struct name, so different `primaryType`
+//! values produce different EIP-712 digests even with identical field
+//! values.
+//!
+//! Server-authoritative model: the client posts the full EIP-712 typed data
+//! the wallet signed. The server validates the domain, the schema for the
+//! claimed `primaryType`, and the timestamp window before doing any
+//! crypto — then recovers the signer from the digest and confirms it
+//! matches the `address` field inside the typed data.
+//!
+//! The wallet UI determines the digest. If the client builds different
+//! bytes than the server expects, the digest differs and recovery fails —
+//! no out-of-band type-hash check needed beyond pinning the schema.
+//!
+//! ## Why this lives in `lit-billing-core`
+//!
+//! Both `lit-api-server` (account-management endpoints) and `lit-payments`
+//! (the `BillingAuth` Rocket guard) need this verifier. Previously the
+//! verifier lived only on `lit-api-server`, and `lit-payments` reached it
+//! via an HTTP hop (`POST /internal/verify_wallet_auth`). The hop was a
+//! pure-function call dressed in network plumbing; folding the verifier
+//! into the shared crate lets `lit-payments` invoke it directly.
+//!
+//! The `chain_id` is supplied per call — the server determines it from its
+//! own config — so this module has no global state of its own. That keeps
+//! the test surface trivial and removes the lit-api-server `OnceLock`
+//! dependency that the original code carried.
+
+use std::collections::BTreeMap;
+
+use alloy_dyn_abi::TypedData;
+use alloy_primitives::{Address, Signature, U256};
+
+/// ±5-minute window on `issuedAt`; the only replay protection (no nonce
+/// store). Worst-case replay on the unauthenticated mint endpoints just
+/// produces an extra unattached PKP — the bytes returned are equivalent to
+/// a freshly generated keypair until the admin wallet calls the on-chain
+/// follow-ups, so compute cost only.
+pub const TIMESTAMP_SKEW_SECONDS: i64 = 300;
+
+/// Hard cap on the JSON-serialised typed-data payload to bound the work an
+/// unauthenticated caller can force the server to do before any cheap
+/// reject. 4 KiB is far above any legitimate payload — the canonical shape
+/// is well under 1 KiB.
+pub const MAX_TYPED_DATA_LEN: usize = 4096;
+
+/// Stable EIP-712 domain `name` for all ChainSecured signatures. Wallet UIs
+/// surface this; do not change without a coordinated client release.
+pub const EIP712_DOMAIN_NAME: &str = "Lit ChainSecured";
+
+/// Domain `version`. Bump only when the typed-data schema changes in a way
+/// that requires re-signing — wallets refuse to interpret a v2 signature
+/// against a v1 domain (and vice versa).
+pub const EIP712_DOMAIN_VERSION: &str = "1";
+
+/// Canonical primary types per ChainSecured flow. Each endpoint pins one
+/// of these and rejects signatures whose `primaryType` doesn't match — the
+/// type hash commits to the struct name, so a signature minted for one
+/// flow cannot recover when the server hashes the same bytes against a
+/// different `primaryType`.
+pub const PRIMARY_TYPE_CREATE_WALLET: &str = "CreateWallet";
+pub const PRIMARY_TYPE_CONVERT_ACCOUNT: &str = "ConvertAccount";
+pub const PRIMARY_TYPE_ADD_USAGE_API_KEY: &str = "AddUsageApiKey";
+pub const PRIMARY_TYPE_BILLING_AUTH: &str = "BillingAuth";
+
+/// Failure modes from [`verify_eip712_signature`]. Two flavours:
+/// - [`Eip712Error::BadRequest`] — caller-side problem (bad payload,
+///   wrong domain/chain/schema, stale/future timestamp, signature mismatch).
+///   Maps to 400/401 on the HTTP edge.
+/// - [`Eip712Error::Internal`] — server-side problem computing the digest.
+///   Maps to 500.
+///
+/// Each variant carries the human-friendly summary as a separate field so
+/// callers can surface a stable message while logging the wrapped error
+/// detail; that's how the original lit-api-server `ApiStatus`-based shape
+/// worked and what the dashboard depends on.
+#[derive(Debug, thiserror::Error)]
+pub enum Eip712Error {
+    #[error("{summary}: {detail}")]
+    BadRequest { summary: String, detail: String },
+    #[error("{summary}: {detail}")]
+    Internal { summary: String, detail: String },
+}
+
+impl Eip712Error {
+    fn bad(detail: impl Into<String>, summary: impl Into<String>) -> Self {
+        Self::BadRequest {
+            summary: summary.into(),
+            detail: detail.into(),
+        }
+    }
+    fn internal(detail: impl Into<String>, summary: impl Into<String>) -> Self {
+        Self::Internal {
+            summary: summary.into(),
+            detail: detail.into(),
+        }
+    }
+    /// Whether this error should be treated as a 4xx (caller fault) vs 5xx
+    /// (server fault). Callers may want to translate it into their own
+    /// response type.
+    pub fn is_bad_request(&self) -> bool {
+        matches!(self, Self::BadRequest { .. })
+    }
+    /// Stable summary string — the second argument the original ApiStatus
+    /// helper took. Suitable for surfacing to the dashboard. The detail
+    /// field is the wrapped error and is meant for server logs.
+    pub fn summary(&self) -> &str {
+        match self {
+            Self::BadRequest { summary, .. } | Self::Internal { summary, .. } => summary,
+        }
+    }
+}
+
+/// One field of an EIP-712 type declaration. Mirrors the wire shape
+/// (`{"name": "...", "type": "..."}`) so we can validate the client-supplied
+/// `types` map byte-for-byte against the canonical schema. We don't use
+/// alloy's `PropertyDef` directly here because comparison ergonomics with
+/// string literals are simpler with a plain struct, and the JSON shape we
+/// validate against is fixed.
+///
+/// `deny_unknown_fields` is load-bearing: a phishing dApp could otherwise
+/// embed a decoy field like `{ "name": "address", "type": "address",
+/// "label": "Approve $500" }`. The extra key doesn't change the EIP-712
+/// type hash (only `name` + `type` are encoded), so the digest still
+/// recovers, but some wallet UIs surface the unknown metadata to the
+/// user. The schema-equality check below sees only the canonical fields
+/// and would silently pass without this guard.
+#[derive(Debug, PartialEq, Eq, serde::Deserialize, serde::Serialize)]
+#[serde(deny_unknown_fields)]
+struct Eip712FieldDef {
+    name: String,
+    #[serde(rename = "type")]
+    ty: String,
+}
+
+impl Eip712FieldDef {
+    fn new(name: &str, ty: &str) -> Self {
+        Self {
+            name: name.to_string(),
+            ty: ty.to_string(),
+        }
+    }
+}
+
+/// Just the slice of the JSON we need to validate the schema. Alloy's
+/// `TypedData::resolver` is opaque (private fields), so we deserialize the
+/// `types` and `primaryType` keys into our own view and run schema checks
+/// against that. The same JSON is also fed to `alloy_dyn_abi::TypedData`
+/// to compute the EIP-712 digest — serde_json yields a deterministic logical
+/// value, so both parses see the same fields.
+/// `deny_unknown_fields` on the top-level view: a phishing payload could
+/// otherwise smuggle in an extra top-level key (e.g. `displayMessage`) that
+/// some wallet UIs surface alongside the signed struct. The four canonical
+/// keys are exhaustive for what this server validates.
+#[derive(Debug, serde::Deserialize)]
+#[serde(deny_unknown_fields)]
+struct TypedDataSchemaView {
+    types: BTreeMap<String, Vec<Eip712FieldDef>>,
+    #[serde(rename = "primaryType")]
+    primary_type: String,
+    domain: DomainView,
+    message: BTreeMap<String, serde_json::Value>,
+}
+
+/// `deny_unknown_fields` is load-bearing on the domain too — the goal is
+/// strict anti-phishing pinning to exactly `(name, version, chainId)`. The
+/// explicit `verifying_contract`/`salt` rejects below still trip if those
+/// fields are present (so the error message stays specific for the common
+/// case), but anything *else* (e.g. `displayName`) is rejected here.
+#[derive(Debug, serde::Deserialize)]
+#[serde(deny_unknown_fields)]
+struct DomainView {
+    name: Option<String>,
+    version: Option<String>,
+    #[serde(rename = "chainId")]
+    chain_id: Option<serde_json::Value>,
+    #[serde(default, rename = "verifyingContract")]
+    verifying_contract: Option<serde_json::Value>,
+    #[serde(default)]
+    salt: Option<serde_json::Value>,
+}
+
+/// The four flows share an identical message struct: `(address, issuedAt)`.
+/// Field declaration order is part of the EIP-712 type hash — clients must
+/// declare the fields in this order, and any reordering is rejected by
+/// `validate_type_schema`.
+fn payload_field_schema() -> [Eip712FieldDef; 2] {
+    [
+        Eip712FieldDef::new("address", "address"),
+        Eip712FieldDef::new("issuedAt", "uint256"),
+    ]
+}
+
+/// Canonical `EIP712Domain` field declaration. We use the (name, version,
+/// chainId) subset — no `verifyingContract` (this server is not an
+/// on-chain contract) and no `salt`. Clients must declare these in the
+/// same order; any other shape is rejected.
+fn domain_field_schema() -> [Eip712FieldDef; 3] {
+    [
+        Eip712FieldDef::new("name", "string"),
+        Eip712FieldDef::new("version", "string"),
+        Eip712FieldDef::new("chainId", "uint256"),
+    ]
+}
+
+fn is_known_primary_type(s: &str) -> bool {
+    matches!(
+        s,
+        PRIMARY_TYPE_CREATE_WALLET
+            | PRIMARY_TYPE_CONVERT_ACCOUNT
+            | PRIMARY_TYPE_ADD_USAGE_API_KEY
+            | PRIMARY_TYPE_BILLING_AUTH
+    )
+}
+
+/// Build the canonical typed data JSON the server expects for a given flow.
+/// Used by tests to round-trip a wallet signature; not part of the request
+/// path (the request path validates the client-supplied typed data instead
+/// of rebuilding it, so the wallet's view and the server's view are bit-
+/// identical when verification succeeds).
+#[cfg(test)]
+pub(crate) fn build_canonical_typed_data_json(
+    primary_type: &str,
+    address: Address,
+    issued_at: i64,
+    chain_id: u64,
+) -> serde_json::Value {
+    let payload_fields: Vec<serde_json::Value> = payload_field_schema()
+        .iter()
+        .map(|f| serde_json::json!({ "name": f.name, "type": f.ty }))
+        .collect();
+    let domain_fields: Vec<serde_json::Value> = domain_field_schema()
+        .iter()
+        .map(|f| serde_json::json!({ "name": f.name, "type": f.ty }))
+        .collect();
+
+    serde_json::json!({
+        "types": {
+            "EIP712Domain": domain_fields,
+            primary_type: payload_fields,
+        },
+        "primaryType": primary_type,
+        "domain": {
+            "name": EIP712_DOMAIN_NAME,
+            "version": EIP712_DOMAIN_VERSION,
+            // chainId is stringified to match the wire shape JS wallets use.
+            "chainId": chain_id.to_string(),
+        },
+        "message": {
+            "address": format!("{:#x}", address),
+            "issuedAt": issued_at.to_string(),
+        },
+    })
+}
+
+/// Verify an EIP-712 typed-data + signature pair for a specific ChainSecured
+/// flow. Returns the recovered wallet address on success.
+///
+/// `expected_chain_id` is the server's notion of which chain a valid
+/// signature must commit to (e.g. Anvil 175188, mainnet 1). Each caller
+/// reads it from its own config — `lit-api-server` from `GLOBAL_NODE_CONFIG`,
+/// `lit-payments` from its `Config.chain_id` env var.
+///
+/// Cheap rejects (length cap, JSON parse, schema match, domain match,
+/// primary-type match, timestamp window) run before the expensive ECDSA
+/// recovery so junk traffic with stale, wrong-purpose, or wrong-chain
+/// payloads is dropped without doing crypto.
+pub fn verify_eip712_signature(
+    typed_data_json: &serde_json::Value,
+    signature_hex: &str,
+    expected_primary_type: &str,
+    expected_chain_id: u64,
+) -> Result<Address, Eip712Error> {
+    // Length cap before parsing — bound the work an unauthenticated caller
+    // can force the server to do before we hit anything expensive.
+    let serialized = serde_json::to_string(typed_data_json)
+        .map_err(|e| Eip712Error::bad(e.to_string(), "typed_data not serializable"))?;
+    if serialized.len() > MAX_TYPED_DATA_LEN {
+        return Err(Eip712Error::bad(
+            format!("typed_data exceeds {MAX_TYPED_DATA_LEN}-byte cap"),
+            "typed_data too large",
+        ));
+    }
+
+    // Accept both wire shapes for the typed-data payload: the object form
+    // (what viem and ethers.js v6 send) and the stringified form (what
+    // `eth_signTypedData_v4` returns from MetaMask, which some clients
+    // forward as-is). Both ethers' and alloy's `TypedData` deserializers
+    // accept both shapes, so the prior verifier handled both; rejecting
+    // the stringified form would silently break those clients.
+    let typed_data_json = normalize_typed_data_input(typed_data_json)?;
+
+    let view: TypedDataSchemaView =
+        serde_json::from_value(typed_data_json.clone()).map_err(|e| {
+            Eip712Error::bad(
+                e.to_string(),
+                "typed_data is not a valid EIP-712 typed-data object",
+            )
+        })?;
+
+    if view.primary_type != expected_primary_type {
+        return Err(Eip712Error::bad(
+            format!(
+                "primaryType mismatch: typed_data says {:?}, this endpoint expects {:?}",
+                view.primary_type, expected_primary_type
+            ),
+            "primaryType mismatch — signature was minted for a different flow",
+        ));
+    }
+    if !is_known_primary_type(&view.primary_type) {
+        // Defence in depth — `expected_primary_type` is always one of our
+        // constants, but pin the wider invariant explicitly so adding a new
+        // primary type forces us to update `is_known_primary_type` too.
+        return Err(Eip712Error::bad(
+            format!("Unknown primaryType {:?}", view.primary_type),
+            "Unknown primaryType",
+        ));
+    }
+
+    validate_domain(&view.domain, expected_chain_id)?;
+    validate_type_schema(&view)?;
+    let address = extract_address(&view)?;
+    let issued_at = extract_issued_at(&view)?;
+    validate_timestamp(issued_at)?;
+
+    // Parse a second time into alloy's TypedData to compute the EIP-712
+    // digest. Two parses of the same JSON yield the same logical value.
+    let typed_data: TypedData = serde_json::from_value(typed_data_json).map_err(|e| {
+        Eip712Error::bad(
+            e.to_string(),
+            "typed_data is not a valid EIP-712 typed-data object",
+        )
+    })?;
+    let digest = typed_data.eip712_signing_hash().map_err(|e| {
+        Eip712Error::internal(
+            format!("eip712_signing_hash failed: {e}"),
+            "eip712_signing_hash failed",
+        )
+    })?;
+    let sig: Signature = signature_hex
+        .trim()
+        .parse()
+        .map_err(|e: alloy_primitives::SignatureError| {
+            Eip712Error::bad(e.to_string(), "Invalid signature hex")
+        })?;
+    let recovered = sig
+        .recover_address_from_prehash(&digest)
+        .map_err(|e| Eip712Error::bad(e.to_string(), "Signature recovery failed"))?;
+
+    if recovered != address {
+        return Err(Eip712Error::bad(
+            "Signature does not match claimed address",
+            "Signature does not match claimed address",
+        ));
+    }
+    Ok(address)
+}
+
+fn validate_domain(domain: &DomainView, expected_chain_id: u64) -> Result<(), Eip712Error> {
+    if domain.name.as_deref() != Some(EIP712_DOMAIN_NAME) {
+        return Err(Eip712Error::bad(
+            format!(
+                "Domain name mismatch: typed_data says {:?}, expected {:?}",
+                domain.name, EIP712_DOMAIN_NAME
+            ),
+            "Domain name mismatch",
+        ));
+    }
+    if domain.version.as_deref() != Some(EIP712_DOMAIN_VERSION) {
+        return Err(Eip712Error::bad(
+            format!(
+                "Domain version mismatch: typed_data says {:?}, expected {:?}",
+                domain.version, EIP712_DOMAIN_VERSION
+            ),
+            "Domain version mismatch",
+        ));
+    }
+    let expected_chain_id = U256::from(expected_chain_id);
+    let actual_chain_id = domain
+        .chain_id
+        .as_ref()
+        .map(|v| parse_u256_loose(v, "chainId"))
+        .transpose()
+        .map_err(|e| Eip712Error::bad(e.to_string(), "Invalid chainId in domain"))?;
+    if actual_chain_id != Some(expected_chain_id) {
+        return Err(Eip712Error::bad(
+            format!(
+                "Chain ID mismatch: typed_data says {:?}, server is on {}",
+                domain.chain_id, expected_chain_id
+            ),
+            "Chain ID mismatch",
+        ));
+    }
+    // Reject extra domain fields. A phishing site could otherwise include
+    // a verifyingContract or salt the user doesn't notice and bind the same
+    // bytes to a different domain than the wallet displays. We strictly
+    // pin the domain to (name, version, chainId).
+    if domain.verifying_contract.is_some() || domain.salt.is_some() {
+        return Err(Eip712Error::bad(
+            "Unexpected domain fields",
+            "Domain must contain only name, version, chainId",
+        ));
+    }
+    Ok(())
+}
+
+/// Parse a `uint256` JSON value (chainId, issuedAt, …) as either a numeric
+/// string ("175188"), a hex string ("0x2ac14"), or a JSON number. Matches
+/// what JS wallets, viem, and ethers.js produce across versions; the
+/// EIP-712 spec doesn't pin a single wire shape for uint256.
+fn parse_u256_loose(v: &serde_json::Value, field: &str) -> Result<U256, anyhow::Error> {
+    match v {
+        serde_json::Value::String(s) => {
+            let s = s.trim();
+            if let Some(rest) = s.strip_prefix("0x").or_else(|| s.strip_prefix("0X")) {
+                U256::from_str_radix(rest, 16)
+                    .map_err(|e| anyhow::anyhow!("invalid hex {field}: {e}"))
+            } else {
+                s.parse::<U256>()
+                    .map_err(|e| anyhow::anyhow!("invalid decimal {field}: {e}"))
+            }
+        }
+        serde_json::Value::Number(n) => {
+            let u = n
+                .as_u64()
+                .ok_or_else(|| anyhow::anyhow!("{field} must be a non-negative integer"))?;
+            Ok(U256::from(u))
+        }
+        _ => Err(anyhow::anyhow!("{field} must be a string or number")),
+    }
+}
+
+/// Accept both wire shapes for the typed-data payload:
+///   1. The object form: `req.typed_data = { types: {...}, ... }`
+///   2. The stringified form: `req.typed_data = "{ \"types\": ... }"`
+///
+/// `eth_signTypedData_v4` returns a JSON string; some clients forward it
+/// verbatim as the `typed_data` field. Both ethers' `TypedData` and
+/// alloy's `TypedData` deserializers accept both shapes, so the prior
+/// verifier silently handled both and we preserve that here.
+fn normalize_typed_data_input(
+    typed_data_json: &serde_json::Value,
+) -> Result<serde_json::Value, Eip712Error> {
+    match typed_data_json {
+        serde_json::Value::String(s) => serde_json::from_str(s).map_err(|e| {
+            Eip712Error::bad(e.to_string(), "typed_data is a string but not valid JSON")
+        }),
+        other => Ok(other.clone()),
+    }
+}
+
+fn validate_type_schema(view: &TypedDataSchemaView) -> Result<(), Eip712Error> {
+    let expected_payload = payload_field_schema();
+    let primary_decl = view.types.get(&view.primary_type).ok_or_else(|| {
+        Eip712Error::bad(
+            format!("Missing types[{}]", view.primary_type),
+            "Missing primaryType in types",
+        )
+    })?;
+    if primary_decl.as_slice() != expected_payload.as_slice() {
+        return Err(Eip712Error::bad(
+            format!("primaryType field schema mismatch — expected {expected_payload:?}"),
+            "primaryType field schema mismatch",
+        ));
+    }
+
+    let expected_domain = domain_field_schema();
+    let domain_decl = view.types.get("EIP712Domain").ok_or_else(|| {
+        Eip712Error::bad(
+            "Missing types[EIP712Domain]",
+            "Missing EIP712Domain in types",
+        )
+    })?;
+    if domain_decl.as_slice() != expected_domain.as_slice() {
+        return Err(Eip712Error::bad(
+            format!("EIP712Domain field schema mismatch — expected {expected_domain:?}"),
+            "EIP712Domain field schema mismatch",
+        ));
+    }
+
+    // Reject any extra type definitions the client tried to smuggle in.
+    // We don't use them, but a wallet might display them (depending on
+    // implementation) and the user could be misled into thinking they
+    // signed something different.
+    if view.types.len() != 2 {
+        return Err(Eip712Error::bad(
+            format!(
+                "Unexpected type definitions — expected only EIP712Domain and {}",
+                view.primary_type
+            ),
+            "Unexpected type definitions",
+        ));
+    }
+
+    // Strict check on `message` content. The EIP-712 type hash only commits
+    // to the fields declared in `types[primaryType]`, so extras in `message`
+    // don't change the digest — but a phishing dApp could still embed
+    // decoy fields (e.g. `intent: "Subscribe to weekly briefing"`) that
+    // some wallet UIs render alongside the canonical fields, fooling the
+    // user into signing what looks like a different action. Pin the
+    // message to exactly the canonical (address, issuedAt) pair.
+    if view.message.len() != 2
+        || !view.message.contains_key("address")
+        || !view.message.contains_key("issuedAt")
+    {
+        return Err(Eip712Error::bad(
+            format!(
+                "message must contain exactly `address` and `issuedAt`, got {:?}",
+                view.message.keys().collect::<Vec<_>>()
+            ),
+            "message contains unexpected fields",
+        ));
+    }
+    Ok(())
+}
+
+fn extract_address(view: &TypedDataSchemaView) -> Result<Address, Eip712Error> {
+    let address_val = view
+        .message
+        .get("address")
+        .ok_or_else(|| Eip712Error::bad("Missing message.address", "Missing message.address"))?;
+    let address: Address = serde_json::from_value(address_val.clone())
+        .map_err(|e| Eip712Error::bad(e.to_string(), "Invalid message.address"))?;
+    Ok(address)
+}
+
+fn extract_issued_at(view: &TypedDataSchemaView) -> Result<i64, Eip712Error> {
+    let issued_val = view
+        .message
+        .get("issuedAt")
+        .ok_or_else(|| Eip712Error::bad("Missing message.issuedAt", "Missing message.issuedAt"))?;
+    // The field is declared `uint256` in the schema. JS wallets generally
+    // emit it as a decimal string, but alloy and ethers both accept hex
+    // strings and JSON numbers for uint256 — and the digest the wallet
+    // signed is computed from whichever wire shape they used. Mirror that
+    // here so we don't reject pre-recovery on a payload that would hash
+    // correctly. Then bound to i64::MAX so the ±skew arithmetic below
+    // can't overflow.
+    let n = parse_u256_loose(issued_val, "issuedAt")
+        .map_err(|e| Eip712Error::bad(e.to_string(), "issuedAt not a uint256 value"))?;
+    let i64_max_as_u256 = U256::from(i64::MAX as u64);
+    if n > i64_max_as_u256 {
+        return Err(Eip712Error::bad(
+            format!("issuedAt exceeds i64::MAX ({})", i64::MAX),
+            "issuedAt out of range",
+        ));
+    }
+    // Safe: bounded to i64::MAX above.
+    let as_u64: u64 = n.to::<u64>();
+    Ok(as_u64 as i64)
+}
+
+fn validate_timestamp(issued_at: i64) -> Result<(), Eip712Error> {
+    let now = std::time::SystemTime::now()
+        .duration_since(std::time::UNIX_EPOCH)
+        .map_err(|e| Eip712Error::internal(e.to_string(), "System clock is before the Unix epoch"))?
+        .as_secs() as i64;
+    // i128 subtraction: two i64s can never overflow into i128, and i128::abs
+    // is safe for any value produced by an i64 subtraction. Without this, a
+    // crafted issuedAt of i64::MIN would wrap `(now - issued_at).abs()` back
+    // to a small/negative value and bypass the skew check in release builds.
+    let delta = (now as i128) - (issued_at as i128);
+    if delta.abs() > TIMESTAMP_SKEW_SECONDS as i128 {
+        return Err(Eip712Error::bad(
+            format!(
+                "Issued At {issued_at} is outside the ±{TIMESTAMP_SKEW_SECONDS}s window from now ({now})"
+            ),
+            "Signed message timestamp is too old or too far in the future",
+        ));
+    }
+    Ok(())
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use alloy_dyn_abi::TypedData;
+    use alloy_signer::SignerSync;
+    use alloy_signer_local::PrivateKeySigner;
+
+    /// All tests share the Anvil chain id (matches the original
+    /// lit-api-server fixture). The verifier takes chain_id as a parameter
+    /// now, so there's no `OnceLock` to seed.
+    const TEST_CHAIN_ID: u64 = 175188;
+
+    fn now_secs() -> i64 {
+        std::time::SystemTime::now()
+            .duration_since(std::time::UNIX_EPOCH)
+            .unwrap()
+            .as_secs() as i64
+    }
+
+    /// Builds canonical typed data, signs it with `wallet`, returns
+    /// `(typed_data_json, signature_hex)` so tests can hand them to the
+    /// verifier in the same shape an HTTP request would.
+    fn sign_canonical(
+        wallet: &PrivateKeySigner,
+        primary_type: &str,
+        issued_at: i64,
+        chain_id: u64,
+    ) -> (serde_json::Value, String) {
+        let json =
+            build_canonical_typed_data_json(primary_type, wallet.address(), issued_at, chain_id);
+        let typed_data: TypedData = serde_json::from_value(json.clone()).unwrap();
+        let digest = typed_data.eip712_signing_hash().unwrap();
+        let sig = wallet.sign_hash_sync(&digest).unwrap();
+        (json, format!("0x{}", hex::encode(sig.as_bytes())))
+    }
+
+    #[test]
+    fn happy_path_create_wallet() {
+        let wallet = PrivateKeySigner::random();
+        let (typed, sig) =
+            sign_canonical(&wallet, PRIMARY_TYPE_CREATE_WALLET, now_secs(), TEST_CHAIN_ID);
+        let recovered =
+            verify_eip712_signature(&typed, &sig, PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+                .unwrap();
+        assert_eq!(recovered, wallet.address());
+    }
+
+    #[test]
+    fn happy_path_all_primary_types() {
+        for primary in [
+            PRIMARY_TYPE_CREATE_WALLET,
+            PRIMARY_TYPE_CONVERT_ACCOUNT,
+            PRIMARY_TYPE_ADD_USAGE_API_KEY,
+            PRIMARY_TYPE_BILLING_AUTH,
+        ] {
+            let wallet = PrivateKeySigner::random();
+            let (typed, sig) = sign_canonical(&wallet, primary, now_secs(), TEST_CHAIN_ID);
+            let recovered = verify_eip712_signature(&typed, &sig, primary, TEST_CHAIN_ID)
+                .unwrap_or_else(|e| panic!("expected {primary} to verify, got: {e:?}"));
+            assert_eq!(recovered, wallet.address(), "{primary} address mismatch");
+        }
+    }
+
+    /// The core CPL-286 promise: a signature minted for one flow must not
+    /// recover when the server hashes the typed data against a different
+    /// primaryType. This is rejected at the schema layer (we check
+    /// `expected_primary_type` before recovery), but the deeper guarantee
+    /// is that even if the schema check were bypassed, the digest would
+    /// differ. Both layers are exercised here.
+    #[test]
+    fn cross_flow_replay_rejected_at_primary_type_check() {
+        let wallet = PrivateKeySigner::random();
+        // User signs a CreateWallet typed payload.
+        let (typed, sig) =
+            sign_canonical(&wallet, PRIMARY_TYPE_CREATE_WALLET, now_secs(), TEST_CHAIN_ID);
+        // Attacker forwards it to the AddUsageApiKey endpoint as-is.
+        let err =
+            verify_eip712_signature(&typed, &sig, PRIMARY_TYPE_ADD_USAGE_API_KEY, TEST_CHAIN_ID)
+                .expect_err("must reject — primaryType mismatch");
+        let msg = format!("{err}");
+        assert!(
+            msg.contains("primaryType") || msg.contains("primary_type"),
+            "unexpected error: {msg}",
+        );
+    }
+
+    /// Even if the attacker rewrites the `primaryType` field after the fact
+    /// (so it matches the target endpoint), the signature will not recover
+    /// because the EIP-712 digest commits to the type hash.
+    #[test]
+    fn cross_flow_replay_rejected_at_recovery_when_primary_type_rewritten() {
+        let wallet = PrivateKeySigner::random();
+        // User signs CreateWallet.
+        let (mut typed, sig) =
+            sign_canonical(&wallet, PRIMARY_TYPE_CREATE_WALLET, now_secs(), TEST_CHAIN_ID);
+        // Attacker rewrites primaryType + types map to look like AddUsageApiKey.
+        typed["primaryType"] = serde_json::json!(PRIMARY_TYPE_ADD_USAGE_API_KEY);
+        let payload_fields = serde_json::json!([
+            { "name": "address", "type": "address" },
+            { "name": "issuedAt", "type": "uint256" },
+        ]);
+        typed["types"]
+            .as_object_mut()
+            .unwrap()
+            .remove(PRIMARY_TYPE_CREATE_WALLET);
+        typed["types"]
+            .as_object_mut()
+            .unwrap()
+            .insert(PRIMARY_TYPE_ADD_USAGE_API_KEY.to_string(), payload_fields);
+        let err =
+            verify_eip712_signature(&typed, &sig, PRIMARY_TYPE_ADD_USAGE_API_KEY, TEST_CHAIN_ID)
+                .expect_err("must reject — digest doesn't match the rewritten type hash");
+        let msg = format!("{err}");
+        assert!(
+            msg.contains("Signature does not match") || msg.contains("Signature recovery failed"),
+            "unexpected error: {msg}",
+        );
+    }
+
+    #[test]
+    fn rejects_chain_id_mismatch() {
+        let wallet = PrivateKeySigner::random();
+        let (typed, sig) = sign_canonical(
+            &wallet,
+            PRIMARY_TYPE_CREATE_WALLET,
+            now_secs(),
+            TEST_CHAIN_ID.wrapping_add(1),
+        );
+        let err =
+            verify_eip712_signature(&typed, &sig, PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+                .expect_err("must reject — wrong chainId");
+        assert!(format!("{err}").contains("Chain ID"));
+    }
+
+    #[test]
+    fn rejects_domain_name_mismatch() {
+        let wallet = PrivateKeySigner::random();
+        let (mut typed, sig) =
+            sign_canonical(&wallet, PRIMARY_TYPE_CREATE_WALLET, now_secs(), TEST_CHAIN_ID);
+        typed["domain"]["name"] = serde_json::json!("Not Lit ChainSecured");
+        let err =
+            verify_eip712_signature(&typed, &sig, PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+                .expect_err("must reject — wrong domain name");
+        assert!(format!("{err}").contains("Domain name"));
+    }
+
+    #[test]
+    fn rejects_domain_version_mismatch() {
+        let wallet = PrivateKeySigner::random();
+        let (mut typed, sig) =
+            sign_canonical(&wallet, PRIMARY_TYPE_CREATE_WALLET, now_secs(), TEST_CHAIN_ID);
+        typed["domain"]["version"] = serde_json::json!("99");
+        let err =
+            verify_eip712_signature(&typed, &sig, PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+                .expect_err("must reject — wrong domain version");
+        assert!(format!("{err}").contains("Domain version"));
+    }
+
+    #[test]
+    fn rejects_extra_domain_field_verifying_contract() {
+        let wallet = PrivateKeySigner::random();
+        let (mut typed, sig) =
+            sign_canonical(&wallet, PRIMARY_TYPE_CREATE_WALLET, now_secs(), TEST_CHAIN_ID);
+        typed["domain"]["verifyingContract"] =
+            serde_json::json!("0x1111111111111111111111111111111111111111");
+        let err =
+            verify_eip712_signature(&typed, &sig, PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+                .expect_err("must reject — verifyingContract is not allowed");
+        assert!(format!("{err}").contains("Domain"));
+    }
+
+    #[test]
+    fn rejects_schema_field_reorder() {
+        let wallet = PrivateKeySigner::random();
+        let (mut typed, sig) =
+            sign_canonical(&wallet, PRIMARY_TYPE_CREATE_WALLET, now_secs(), TEST_CHAIN_ID);
+        // Swap field order on the primary type — produces a different type hash.
+        typed["types"][PRIMARY_TYPE_CREATE_WALLET] = serde_json::json!([
+            { "name": "issuedAt", "type": "uint256" },
+            { "name": "address", "type": "address" },
+        ]);
+        let err =
+            verify_eip712_signature(&typed, &sig, PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+                .expect_err("must reject — schema fields reordered");
+        assert!(format!("{err}").contains("schema"));
+    }
+
+    #[test]
+    fn rejects_extra_type_definitions() {
+        let wallet = PrivateKeySigner::random();
+        let (mut typed, sig) =
+            sign_canonical(&wallet, PRIMARY_TYPE_CREATE_WALLET, now_secs(), TEST_CHAIN_ID);
+        // Smuggle in an extra type the wallet UI might display alongside.
+        typed["types"]["TotallyLegit"] = serde_json::json!([
+            { "name": "spookyField", "type": "string" },
+        ]);
+        let err =
+            verify_eip712_signature(&typed, &sig, PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+                .expect_err("must reject — types must contain exactly EIP712Domain + primaryType");
+        assert!(format!("{err}").contains("Unexpected type definitions"));
+    }
+
+    #[test]
+    fn rejects_timestamp_too_old() {
+        let wallet = PrivateKeySigner::random();
+        let stale = now_secs() - TIMESTAMP_SKEW_SECONDS - 1;
+        let (typed, sig) =
+            sign_canonical(&wallet, PRIMARY_TYPE_CREATE_WALLET, stale, TEST_CHAIN_ID);
+        let err =
+            verify_eip712_signature(&typed, &sig, PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+                .expect_err("must reject — issued_at outside skew window");
+        assert!(format!("{err}").contains("timestamp"));
+    }
+
+    #[test]
+    fn rejects_timestamp_too_far_future() {
+        let wallet = PrivateKeySigner::random();
+        let future = now_secs() + TIMESTAMP_SKEW_SECONDS + 3600;
+        let (typed, sig) =
+            sign_canonical(&wallet, PRIMARY_TYPE_CREATE_WALLET, future, TEST_CHAIN_ID);
+        let err =
+            verify_eip712_signature(&typed, &sig, PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+                .expect_err("must reject — issued_at too far in future");
+        assert!(format!("{err}").contains("timestamp"));
+    }
+
+    #[test]
+    fn rejects_i64_min_issued_at() {
+        let typed = serde_json::json!({
+            "types": {
+                "EIP712Domain": [
+                    { "name": "name", "type": "string" },
+                    { "name": "version", "type": "string" },
+                    { "name": "chainId", "type": "uint256" },
+                ],
+                "CreateWallet": [
+                    { "name": "address", "type": "address" },
+                    { "name": "issuedAt", "type": "uint256" },
+                ],
+            },
+            "primaryType": "CreateWallet",
+            "domain": {
+                "name": "Lit ChainSecured",
+                "version": "1",
+                "chainId": TEST_CHAIN_ID.to_string(),
+            },
+            "message": {
+                "address": "0x0000000000000000000000000000000000000000",
+                "issuedAt": i64::MIN.to_string(),
+            }
+        });
+        let err = verify_eip712_signature(&typed, "0x00", PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+            .expect_err("must reject — issued_at is negative, not a valid uint256");
+        assert!(format!("{err}").contains("issuedAt"));
+    }
+
+    #[test]
+    fn rejects_issued_at_overflows_i64() {
+        let huge = "99999999999999999999"; // > i64::MAX
+        let typed = serde_json::json!({
+            "types": {
+                "EIP712Domain": [
+                    { "name": "name", "type": "string" },
+                    { "name": "version", "type": "string" },
+                    { "name": "chainId", "type": "uint256" },
+                ],
+                "CreateWallet": [
+                    { "name": "address", "type": "address" },
+                    { "name": "issuedAt", "type": "uint256" },
+                ],
+            },
+            "primaryType": "CreateWallet",
+            "domain": {
+                "name": "Lit ChainSecured",
+                "version": "1",
+                "chainId": TEST_CHAIN_ID.to_string(),
+            },
+            "message": {
+                "address": "0x0000000000000000000000000000000000000000",
+                "issuedAt": huge,
+            }
+        });
+        let err = verify_eip712_signature(&typed, "0x00", PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+            .expect_err("must reject — issuedAt out of i64 range");
+        assert!(format!("{err}").contains("out of range"));
+    }
+
+    #[test]
+    fn rejects_extra_key_in_field_def() {
+        let typed = serde_json::json!({
+            "types": {
+                "EIP712Domain": [
+                    { "name": "name", "type": "string" },
+                    { "name": "version", "type": "string" },
+                    { "name": "chainId", "type": "uint256" },
+                ],
+                "CreateWallet": [
+                    { "name": "address", "type": "address", "label": "Approve $500" },
+                    { "name": "issuedAt", "type": "uint256" },
+                ],
+            },
+            "primaryType": "CreateWallet",
+            "domain": {
+                "name": "Lit ChainSecured",
+                "version": "1",
+                "chainId": TEST_CHAIN_ID.to_string(),
+            },
+            "message": {
+                "address": "0x0000000000000000000000000000000000000000",
+                "issuedAt": now_secs().to_string(),
+            }
+        });
+        let err = verify_eip712_signature(&typed, "0x00", PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+            .expect_err("must reject — field def has an extra key");
+        assert!(
+            format!("{err}").contains("label") || format!("{err}").contains("unknown field"),
+            "unexpected error: {err}",
+        );
+    }
+
+    #[test]
+    fn rejects_extra_key_in_domain() {
+        let wallet = PrivateKeySigner::random();
+        let (mut typed, sig) =
+            sign_canonical(&wallet, PRIMARY_TYPE_CREATE_WALLET, now_secs(), TEST_CHAIN_ID);
+        typed["domain"]["displayName"] = serde_json::json!("Approve transfer");
+        let err =
+            verify_eip712_signature(&typed, &sig, PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+                .expect_err("must reject — unknown domain key");
+        assert!(
+            format!("{err}").contains("displayName") || format!("{err}").contains("unknown field"),
+            "unexpected error: {err}",
+        );
+    }
+
+    #[test]
+    fn accepts_stringified_typed_data() {
+        let wallet = PrivateKeySigner::random();
+        let (json, sig) =
+            sign_canonical(&wallet, PRIMARY_TYPE_CREATE_WALLET, now_secs(), TEST_CHAIN_ID);
+        let stringified = serde_json::Value::String(serde_json::to_string(&json).unwrap());
+        let recovered = verify_eip712_signature(
+            &stringified,
+            &sig,
+            PRIMARY_TYPE_CREATE_WALLET,
+            TEST_CHAIN_ID,
+        )
+        .expect("stringified typed_data must verify");
+        assert_eq!(recovered, wallet.address());
+    }
+
+    #[test]
+    fn accepts_hex_issued_at() {
+        let wallet = PrivateKeySigner::random();
+        let mut json = build_canonical_typed_data_json(
+            PRIMARY_TYPE_CREATE_WALLET,
+            wallet.address(),
+            now_secs(),
+            TEST_CHAIN_ID,
+        );
+        let now = now_secs();
+        json["message"]["issuedAt"] = serde_json::json!(format!("0x{:x}", now));
+        let typed_data: TypedData = serde_json::from_value(json.clone()).unwrap();
+        let digest = typed_data.eip712_signing_hash().unwrap();
+        let sig = wallet.sign_hash_sync(&digest).unwrap();
+        let sig_hex = format!("0x{}", hex::encode(sig.as_bytes()));
+        let recovered =
+            verify_eip712_signature(&json, &sig_hex, PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+                .expect("hex issuedAt must verify");
+        assert_eq!(recovered, wallet.address());
+    }
+
+    #[test]
+    fn rejects_address_signer_mismatch() {
+        let signer = PrivateKeySigner::random();
+        let other = PrivateKeySigner::random();
+        let json = build_canonical_typed_data_json(
+            PRIMARY_TYPE_CREATE_WALLET,
+            other.address(),
+            now_secs(),
+            TEST_CHAIN_ID,
+        );
+        let typed_data: TypedData = serde_json::from_value(json.clone()).unwrap();
+        let digest = typed_data.eip712_signing_hash().unwrap();
+        let sig = signer.sign_hash_sync(&digest).unwrap();
+        let err = verify_eip712_signature(
+            &json,
+            &format!("0x{}", hex::encode(sig.as_bytes())),
+            PRIMARY_TYPE_CREATE_WALLET,
+            TEST_CHAIN_ID,
+        )
+        .expect_err("must reject — recovered signer ≠ claimed address");
+        assert!(format!("{err}").contains("Signature does not match"));
+    }
+
+    #[test]
+    fn rejects_bad_signature_hex() {
+        let wallet = PrivateKeySigner::random();
+        let (typed, _) =
+            sign_canonical(&wallet, PRIMARY_TYPE_CREATE_WALLET, now_secs(), TEST_CHAIN_ID);
+        let err =
+            verify_eip712_signature(&typed, "0xnothex", PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+                .expect_err("must reject — bad signature hex");
+        assert!(format!("{err}").contains("signature") || format!("{err}").contains("Signature"));
+    }
+
+    #[test]
+    fn rejects_malformed_typed_data_json() {
+        let bogus = serde_json::json!({ "not": "a typed data object" });
+        let err =
+            verify_eip712_signature(&bogus, "0x00", PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+                .expect_err("must reject — JSON does not deserialize as TypedData");
+        assert!(format!("{err}").to_lowercase().contains("typed_data"));
+    }
+
+    #[test]
+    fn rejects_oversized_typed_data() {
+        let mut typed = serde_json::json!({
+            "types": {
+                "EIP712Domain": [
+                    { "name": "name", "type": "string" },
+                    { "name": "version", "type": "string" },
+                    { "name": "chainId", "type": "uint256" },
+                ],
+                "CreateWallet": [
+                    { "name": "address", "type": "address" },
+                    { "name": "issuedAt", "type": "uint256" },
+                ],
+            },
+            "primaryType": "CreateWallet",
+            "domain": {
+                "name": "Lit ChainSecured",
+                "version": "1",
+                "chainId": TEST_CHAIN_ID.to_string(),
+            },
+            "message": {
+                "address": "0x0000000000000000000000000000000000000000",
+                "issuedAt": "0",
+            }
+        });
+        typed["domain"]["name"] = serde_json::Value::String("x".repeat(MAX_TYPED_DATA_LEN + 1));
+        let err = verify_eip712_signature(&typed, "0x00", PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+            .expect_err("must reject — typed_data exceeds size cap");
+        assert!(format!("{err}").contains("too large"));
+    }
+
+    #[test]
+    fn rejects_extra_message_fields() {
+        let wallet = PrivateKeySigner::random();
+        let (mut typed, sig) =
+            sign_canonical(&wallet, PRIMARY_TYPE_CREATE_WALLET, now_secs(), TEST_CHAIN_ID);
+        typed["message"]["decoy"] = serde_json::json!("Authorize $10 subscription");
+        let err =
+            verify_eip712_signature(&typed, &sig, PRIMARY_TYPE_CREATE_WALLET, TEST_CHAIN_ID)
+                .expect_err("must reject — message must contain only address + issuedAt");
+        assert!(format!("{err}").contains("unexpected fields"));
+    }
+}
